requirements.txt
httpx[http2]==0.24.1
aiofiles==23.1.0
duckduckgo-search==2.7.0
pytest==7.4.0
pytest-asyncio==0.21.0
minio==7.1.3
python-dotenv==1.0.0
openai==1.0.0
